refactor: decompose discovery.py into resolvers/search/ subpackage (#118)
Extract 10 provider-specific search modules from the 1539 LOC discovery.py
monolith into resolvers/search/ subpackage:

- _common.py: shared HTTP client, headers, timeouts, text cleaning helpers
- gallica.py, vatican.py, harvard.py, archive_org.py, heidelberg.py,
  cambridge.py, institut.py, ecodices.py, bodleian.py, loc.py
- __init__.py: re-exports all public search functions

discovery.py is now ~178 LOC slim orchestration layer that re-exports
all search_* functions for full backward compatibility (providers.py
getattr lookup, test imports).

Updated 6 test files to patch _common._http_client_cache and
vatican module internals at their new locations.

Closes #118
