+++ tests/test_search_institut_unit.py
@@ -4,6 +4,7 @@ import requests
 
 from universal_iiif_core.http_client import HTTPClient
 from universal_iiif_core.resolvers import discovery
+from universal_iiif_core.resolvers.search import _common as _search_common
 
 
 class _Resp:
@@ -27,7 +28,7 @@ def _patch_http_client(monkeypatch, fake_get):
         def get(self, *args, **kwargs):
             return fake_get(*args, **kwargs)
 
-    monkeypatch.setattr(discovery, "_http_client_cache", _MockClient())
+    monkeypatch.setattr(_search_common, "_http_client_cache", _MockClient())
 
 
 def test_search_institut_blank_input_returns_empty():
