+++ tests/test_search_new_providers_unit.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from universal_iiif_core.http_client import HTTPClient
 from universal_iiif_core.resolvers import discovery
+from universal_iiif_core.resolvers.search import _common as _search_common
 
 
 class _Resp:
@@ -28,7 +29,7 @@ def _patch_http_client(monkeypatch, fake_get):
         def get(self, *args, **kwargs):
             return fake_get(*args, **kwargs)
 
-    monkeypatch.setattr(discovery, "_http_client_cache", _MockClient())
+    monkeypatch.setattr(_search_common, "_http_client_cache", _MockClient())
 
 
 def test_search_loc_maps_loc_item_urls_to_manifests(monkeypatch):
